+++ src/JLD2.jl
@@ -509,6 +509,7 @@ include("global_heaps.jl")
 include("data/type_defs.jl")
 include("data/specialcased_types.jl")
 include("data/number_types.jl")
+include("data/anonymous_functions.jl")
 include("data/custom_serialization.jl")
 include("data/writing_datatypes.jl")
 include("data/reconstructing_datatypes.jl")
